refactor: move curve api validation into its own file
