+++ apps/main/src/dao/entities/user-gauge-vote-next-time.ts
@@ -3,7 +3,8 @@ import type { ChainId } from '@/dao/types/dao.types'
 import { requireLib } from '@ui-kit/features/connect-wallet'
 import type { ChainQuery, UserQuery } from '@ui-kit/lib/model/query'
 import { queryFactory } from '@ui-kit/lib/model/query'
-import { chainValidationGroup, curveApiValidationGroup } from '@ui-kit/lib/model/query/chain-validation'
+import { chainValidationGroup } from '@ui-kit/lib/model/query/chain-validation'
+import { curveApiValidationGroup } from '@ui-kit/lib/model/query/curve-api-validation'
 import { userAddressValidationGroup } from '@ui-kit/lib/model/query/user-address-validation'
 import { createValidationSuite, type FieldsOf } from '@ui-kit/lib/validation'
