+++ apps/main/src/loan/entities/appstats-total-crvusd-supply.ts
@@ -3,7 +3,7 @@ import type { ChainId } from '@/loan/types/loan.types'
 import { requireLib } from '@ui-kit/features/connect-wallet'
 import type { ChainParams, ChainQuery } from '@ui-kit/lib/model/query'
 import { queryFactory } from '@ui-kit/lib/model/query'
-import { curveApiValidationSuite } from '@ui-kit/lib/model/query/chain-validation'
+import { curveApiValidationSuite } from '@ui-kit/lib/model/query/curve-api-validation'
 
 export const { useQuery: useAppStatsTotalCrvusdSupply } = queryFactory({
   queryKey: (params: ChainParams<ChainId>) => ['appStatsTotalCrvusdSupply', { chainId: params.chainId }] as const,
